restructured RAN
