+++ model/basic_layers.py
@@ -13,15 +13,15 @@ class ResidualBlock(nn.Module):
         self.stride = stride
         self.bn1 = nn.BatchNorm2d(input_channels)
         self.relu = nn.ReLU(inplace=True)
-        self.conv1 = nn.Conv2d(input_channels, output_channels/4, 1, 1, bias = False)
-        self.bn2 = nn.BatchNorm2d(output_channels/4)
+        self.conv1 = nn.Conv2d(input_channels, output_channels//4, 1, 1, bias = False)
+        self.bn2 = nn.BatchNorm2d(output_channels//4)
         self.relu = nn.ReLU(inplace=True)
-        self.conv2 = nn.Conv2d(output_channels/4, output_channels/4, 3, stride, padding = 1, bias = False)
-        self.bn3 = nn.BatchNorm2d(output_channels/4)
+        self.conv2 = nn.Conv2d(output_channels//4, output_channels//4, 3, stride, padding = 1, bias = False)
+        self.bn3 = nn.BatchNorm2d(output_channels//4)
         self.relu = nn.ReLU(inplace=True)
-        self.conv3 = nn.Conv2d(output_channels/4, output_channels, 1, 1, bias = False)
+        self.conv3 = nn.Conv2d(output_channels//4, output_channels, 1, 1, bias = False)
         self.conv4 = nn.Conv2d(input_channels, output_channels , 1, stride, bias = False)
-        
+
     def forward(self, x):
         residual = x
         out = self.bn1(x)
